Remove lib from this commit
